Suggestion to addressed PR comment
